agentic_chatbot/backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
sqlmodel>=0.0.8
python-jose[cryptography]>=3.3.0
requests>=2.28.0
google-auth>=2.20.0
google-auth-oauthlib>=1.2.0
psycopg2-binary>=2.9.6
langchain>=0.0.287
qdrant-client>=1.2.0
transformers>=4.30.0
sentence-transformers>=2.2.2
langchain-hub>=0.0.2
